requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
Click==7.0
coverage==4.5.2
Flask==1.0.2
funcsigs==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
more-itertools==6.0.0
numpy==1.16.2
pandas==0.24.1
pathlib2==2.3.3
pluggy==0.9.0
py==1.8.0
pyodbc==4.0.26
pytest==4.3.0
python-dateutil==2.8.0
python-interface==1.5.1
pytz==2018.9
-e git+https://github.com/rustyBilges/rankapp.git@d1c3ba6547c7a856b23659973fe3e88707987c2c#egg=rankapp
six==1.12.0
typing==3.6.6
Werkzeug==0.14.1
